int used for remove decimals

+++ Main.java
@@ -1,6 +1,6 @@
 public class Main {
   public static void main(String[] args) {
-    double result= Math.round(Math.random()*100);
+    int result= (int)Math.round(Math.random()*100);
        System.out.println(result);
   }
 }
